requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
asyncpg==0.30.0
attrs==23.1.0
Authlib==1.6.0
av==10.0.0
babel==2.16.0
beautifulsoup4==4.12.2
bleach==6.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.9
decorator==4.4.2
defusedxml==0.7.1
dlib==19.24.4
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.1.2
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.12
fastjsonschema==2.21.0
fonttools==4.55.0
fqdn==1.5.1
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
imbalanced-learn==0.12.4
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
MouseInfo==0.1.3
moviepy==1.0.3
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.24.3
opencv-python==4.9.0.80
openfoodfacts==2.6.0
openpyxl==3.1.5
outcome==1.2.0
overrides==7.7.0
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
Pillow==9.5.0
platformdirs==4.3.6
pluggy==1.6.0
proglog==0.1.10
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.11.5
pydantic_core==2.33.2
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-vlc==3.0.20123
pytweening==1.2.0
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzbar==0.1.9
pyzmq==24.0.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.11.2
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.65.0
traitlets==5.14.3
trio==0.22.2
trio-websocket==0.10.3
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.0.2
uvicorn==0.34.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xgboost==2.1.3
